Refactor detection of OSs from build numbers in simple cases
Refactored code that dectected OSs that were identified by build number and
where their Extra Update Information was a simple plain text string that
required no formatting. The build numbers and extra information were gathered
into tables and a function was added that looked up the required build number
and returned it along with its extra info. This removed a lot (but not all) the
duplicated code in the unit's initialisation.