Fix bug testing for Win 10 & 11 (& equiv servers)
Uses the new TestBuildNumber function to check the OS build number to
determine the actual OS product for all Windows and Windows Server
products that report v10.0.

This new code doesn't require that InternalBuildNumber has been set.
Previously if InternalBuildNumber was 0, the product defaulted to
Win 10 or Win 10 Server, regardless of the actual OS.