Add new Win10/11 related methods to TPJOSInfo
IsWindows10VersionOrLater - checks if OS is the same or later than a
given Windows 10 "version" such as 1507.

IsWindows11VersionOrLater - checks if OS is the same or later than a
given Windows 11 "version" such as 23H2.

Windows10PlusVersion - returns an identifier that specifies the OS
"version" iff it is Win 10 or 11.

Windows10PlusVersionName - returns name of the OS version iff it is Win
10 or 11.

Fixes #17