Update Win10 code for handling end of support
The HiRev field of TBuildNameMap had been used to detect the last ever
build number of a given Windows 10 version. Following the end of support
for Windows 10 this data is no longer available, so the affected field
was removed.

Removed Win10 version 1903 from the Win10_BuildMap table and handled it
specially to ensure that the version 1909 previews with the version 1903
build number are handled correctly.